feat: add navbar and icon, logo common-components
