--- components/common/index.ts
@@ -1,2 +1,0 @@
-export { default as HStack } from "./stacks/HStack";
-export { default as VStack } from "./stacks/VStack";
