--- pages/index.tsx
@@ -1,13 +1,0 @@
-import type { NextPage } from "next";
-import React from "react";
-import styled from "@emotion/styled";
-
-const Home: NextPage = () => {
-  return <PageContainer>취향</PageContainer>;
-};
-
-const PageContainer = styled.section`
-  min-height: 10vh;
-`;
-
-export default Home;
