--- pages/sign-up/index.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const SignUpPage = () => {
-  return <div>회원가입 페이지</div>;
-};
-
-export default SignUpPage;
