+++ public/assets/fonts/fonts.css
@@ -1,0 +1,27 @@
+@font-face {
+  font-family: "KoPubWorld_light";
+  src: url("./KoPubWorld\ Batang\ Light.woff") format("woff");
+  font-weight: normal;
+  font-style: normal;
+}
+
+@font-face {
+  font-family: "KoPubWorld_medium";
+  src: url("./KoPubWorld\ Batang\ Medium.woff") format("woff");
+  font-weight: normal;
+  font-style: normal;
+}
+
+@font-face {
+  font-family: "KoPubWorld_bold";
+  src: url("./KoPubWorld\ Batang\ Bold.woff") format("woff");
+  font-weight: normal;
+  font-style: normal;
+}
+
+@font-face {
+  font-family: "NotoSansKR_regular";
+  src: url("./NotoSansKR-Regular.woff") format("font-woff");
+  font-weight: normal;
+  font-style: normal;
+}
