+++ src/constants/path.ts
@@ -1,0 +1,3 @@
+export enum BaseUrl {
+  Base = "/",
+}
