+++ src/pages/_document.tsx
@@ -1,0 +1,20 @@
+import { Html, Head, Main, NextScript } from "next/document";
+
+const Document = () => {
+  return (
+    <Html>
+      <Head>
+        <link
+          href="https://fonts.googleapis.com/icon?family=Material+Icons|Material+Icons+Outlined|Material+Icons+Two+Tone|Material+Icons+Round|Material+Icons+Sharp"
+          rel="stylesheet"
+        />
+      </Head>
+      <body>
+        <Main />
+        <NextScript />
+      </body>
+    </Html>
+  );
+};
+
+export default Document;
