+++ src/pages/detail/[productDetail]/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ProductDetailPage = () => {
+  return <div>제품 상세 페이지</div>;
+};
+
+export default ProductDetailPage;
