+++ src/pages/index.tsx
@@ -1,0 +1,8 @@
+import type { NextPage } from "next";
+import React from "react";
+
+const Home: NextPage = () => {
+  return <div>취향</div>;
+};
+
+export default Home;
