+++ src/pages/login/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const LoginPage = () => {
+  return <div>로그인 페이지</div>;
+};
+
+export default LoginPage;
