+++ src/pages/my-page/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const MyPage = () => {
+  return <div>마이페이지</div>;
+};
+
+export default MyPage;
