+++ src/pages/search-result/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const SearchResultPage = () => {
+  return <div>검색 결과 페이지</div>;
+};
+
+export default SearchResultPage;
