--- tailwind.config.js
@@ -1,12 +1,0 @@
-/**
- * @type {import('@types/tailwindcss/tailwind-config').TailwindConfig}
- */
-module.exports = {
-  content: [
-    "./node_modules/flowbite-react/**/*.js",
-    "./pages/**/*.{js,jsx,ts,tsx}",
-    "./public/**/*.html",
-  ],
-  plugins: [require("flowbite/plugin")],
-  theme: {},
-};
