+++ tsconfig.json
@@ -18,11 +18,7 @@
     "paths": {
       "@src/*": ["src/*"],
       "@components/*": ["src/components/*"],
-      "@common/*": ["src/common/*"],
       "@common-components": ["src/components/common/index.ts"],
-      // "@hooks/*": ["src/common/hooks/*"],
-      // "@utils/*": ["src/common/utils/*"],
-      // "@constants/*": ["src/common/constants/*"],
       "@styles/*": ["src/styles/*"],
       "@public/*": ["public/*"],
       "@pages/*": ["src/pages/*"]
